+++ src/XrmDefinitelyTyped/Resources/Extensions/xrm_ext_6-.d.ts
@@ -64,7 +64,7 @@ declare namespace Xrm {
   }
 
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  interface UiModule<T extends TabCollectionBase, U extends ControlCollectionBase> {
+  interface UiModule<T extends TabCollectionBase, C extends ControlCollectionBase, Q extends QuickViewFormCollectionBase = QuickViewFormCollectionBase> {
     /**
      * Use this method to remove form level notifications.
      *