+++ test/src/tests/XrmQuery/web/result-type/retrieve.ts
@@ -6,9 +6,11 @@ import { suite, test, slow, timeout, skip, only } from "@testdeck/mocha";
 class Web_Retrieve_ResultTypeCheck {
 
     contactId: string;
+    responseId: string;
 
     before() {
         this.contactId = "CONTACT_ID";
+        this.responseId = "RESPONSE_ID";
     }
 
     @test
@@ -39,4 +41,10 @@ class Web_Retrieve_ResultTypeCheck {
             .execute(x => x.birthdate_formatted)
     }
 
+    @test
+    "select, attribute name matches entity name"() {
+        XrmQuery.retrieve(x => x.dg_responses, this.responseId)
+            .select(x => [x.dg_response])
+            .execute(x => x.dg_response)
+    }
 }